last/requirements.txt
django-cors-headers
django-extensions
django-sslserver
Pillow
torch
